WIP: Current working state creating LB infra (7)
Finish calculation of port forwardings (redirects).

The resulting dict contains the floating ips of the jump hosts. And for each jump host
there is a list of related vms (= vms located in the same az as the jump host) including
their internal ips and an associated port inside the range [port_start, port_end]
which the jump host will use for redirecting traffic to the vm.

The resulting dict might look like this:
        {
          'scs-hm-infra-jh0': {
            'fip': '213.131.230.205',
            'vms': [
              {
                'port': 222,
                'addr': '10.250.3.73',
                'vm_name': 'scs-hm-infra-vm0'
              },
              {
                'port': 223,
                'addr': '10.250.0.8',
                'vm_name': 'scs-hm-infra-vm1'
              }
            ]
          }
        }

Signed-off-by: Katharina Trentau <katharina.trentau@cloudandheat.com>