WIP: Current working state creating LB infra (8)
Change order of steps in order to be able to pass
port redirections as user data (cloud-init) when
the jump hosts are created.

Passing the port redirections this way will prevent
us from logging in to the jump hosts via ssh since
we configure the jump host on its startup.

Signed-off-by: Katharina Trentau <katharina.trentau@cloudandheat.com>