Use alpine instead of busybox for dummy containers; add watch
This allows dummy containers to use a proper "init" process, tini.
It handles SIGINT and SIGTERM, which busybox does not. So sending
SIGTERM via "kubectl delete" or via the Python client exits the dummy
pod immediately, not waiting for timeout or graceperiod as before.

Also adds watches for container creation and deletion, so sleep can be
removed.

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>