Unify pod creation/ping/deletion into one feature; split container steps
This commit merges basic pod tests together: creation of two, ping
between and deletion of both. The Python code in the steps is now split
and can be extended in context of their usage (pod, service, other).

The kubernetes package requirement is replaced from a git-based version
tag to a fixed version from PyPI.

Using the pod name as ping target does not always work, so we fetch the
pod's IP instead. This shall do no more than checking if both pods have
received IPs. (meaning that the cluster's CNI is working)

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>