Rework some container test code (#134)
* Use alpine instead of busybox for dummy containers; add watch

This allows dummy containers to use a proper "init" process, tini.
It handles SIGINT and SIGTERM, which busybox does not. So sending
SIGTERM via "kubectl delete" or via the Python client exits the dummy
pod immediately, not waiting for timeout or graceperiod as before.

Also adds watches for container creation and deletion, so sleep can be
removed.

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>

* Unify pod creation/ping/deletion into one feature; split container steps

This commit merges basic pod tests together: creation of two, ping
between and deletion of both. The Python code in the steps is now split
and can be extended in context of their usage (pod, service, other).

The kubernetes package requirement is replaced from a git-based version
tag to a fixed version from PyPI.

Using the pod name as ping target does not always work, so we fetch the
pod's IP instead. This shall do no more than checking if both pods have
received IPs. (meaning that the cluster's CNI is working)

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>

* Remove unnecessary comments

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>

---------

Signed-off-by: Dominik Pataky <pataky@osb-alliance.com>